Add files via upload
This code analyzes the crime data from the City of Phoenix. It graphs the amount of crime incidents that occurred each day from 2016 to 2018, then it shows a yellow line that represents the dates in which there was a full moon.

Our team concluded that there was no relationship between the lunar cycle and the crime incidents.